implement checking & handling if the id is a valid object id

+++ controllers/ProductControllers.js
@@ -1,5 +1,9 @@
 const Product = require('../models/product_models')
 const asyncHandler = require('express-async-handler')
+const { Types } = require('mongoose')
+
+
+const isValidObjectId = Types.ObjectId.isValid
 
 
 // get all products
@@ -17,11 +21,18 @@ const getProducts = asyncHandler(async (req, res) => {
 const getSpecificProduct = asyncHandler(async (req, res) => {
     try {
         const id  = req.params.id
+        if (!isValidObjectId(id) ) {
+            res.status(400);
+            throw new Error(`Invalid product ID: <${id}> should be 24 characters`)
+        }
         const product = await Product.findById(id)
 
-        if (!product) return res.status(404).json({ message: `Cannot find product with ID in database ${id}`})
-
+        if (!product)  {
+            res.status(404);
+            throw new Error(`Cannot find product with ID <${id}> in database`);
+        }
         res.status(200).json(product)
+
     } catch (error) {
         res.status(500)
         throw new Error(error.message)
@@ -43,11 +54,15 @@ const createProduct = asyncHandler(async (req, res) => {
 const updateProduct = asyncHandler(async (req, res) => {
     try {
         const id = req.params.id
+        if (!isValidObjectId(id) ) {
+            res.status(400);
+            throw new Error(`Invalid product ID: <${id}> should be 24 characters`)
+        }
         const product = await Product.findByIdAndUpdate(id, req.body)
         
         if (!product) {
             res.status(404);
-            throw new Error(`Cannot find product with ID in database ${id}`);
+            throw new Error(`Cannot find product with ID <${id}> in database`);
         }
         
         const updatedProduct = await Product.findById(id)
@@ -62,11 +77,15 @@ const updateProduct = asyncHandler(async (req, res) => {
 const deleteProduct = asyncHandler(async (req, res) => {
     try {
         const { id } = req.params
+        if (!isValidObjectId(id) ) {
+            res.status(400);
+            throw new Error(`Invalid product ID: <${id}> should be 24 characters`)
+        }
         const product = await Product.findByIdAndDelete(id)
         
         if (!product) {
             res.status(404);
-            throw new Error(`Cannot find product with ID in database ${id}`);
+            throw new Error(`Cannot find product with ID <${id}> in database`);
         }
         res.status(200).json(product);
